edited some warnings

+++ FindMedianOfSortedArray.cpp
@@ -4,7 +4,7 @@ double findMedianSortedArrays(vector<int>& nums1, vector<int>& nums2)
 {
 
     vector<int> vec;
-    int nSize = 0;
+    size_t nSize = 0;
     vec.insert(vec.begin(), nums1.begin(), nums1.end());
     vec.insert(vec.end(), nums2.begin(), nums2.end());
